Throw out strtok and its friends

+++ src/lib/string/strlen.c
@@ -8,7 +8,7 @@
 
 #include <string.h>
 
-int strlen(const char * str) {
+size_t strlen(const char * str) {
 	const char *eos = str;
 
 	while (*eos++)
